--- a/liquid-compiler/Cargo.toml
+++ b/liquid-compiler/Cargo.toml
@@ -15,8 +15,9 @@ travis-ci = { repository = "cobalt-org/liquid-rust" }
 appveyor  = { repository = "johannhof/liquid-rust" }
 
 [dependencies]
-regex = "1.0"
-lazy_static = "1.0"
+pest = "2.0"
+pest_derive = "2.0"
+
 # Exposed in API
 liquid-error = { version = "0.16", path = "../liquid-error" }
 liquid-value = { version = "0.17", path = "../liquid-value" }