--- a/liquid-compiler/src/lib.rs
+++ b/liquid-compiler/src/lib.rs
@@ -1,19 +1,16 @@
-#![warn(unused_extern_crates)]
-
-#[macro_use]
-extern crate lazy_static;
 extern crate liquid_error;
 extern crate liquid_interpreter;
 extern crate liquid_value;
-extern crate regex;
+
+extern crate pest;
+#[macro_use]
+extern crate pest_derive;
 
 mod block;
 mod include;
-mod lexer;
 mod options;
 mod parser;
 mod tag;
-mod token;
 
 pub mod error {
     pub use liquid_error::*;
@@ -25,8 +22,6 @@ pub mod value {
 
 pub use block::*;
 pub use include::*;
-pub use lexer::*;
 pub use options::*;
 pub use parser::*;
 pub use tag::*;
-pub use token::*;