--- a/tests/fixtures/input/chained_filters.txt
+++ b/tests/fixtures/input/chained_filters.txt
@@ -1,5 +1,5 @@
 {{foo | replace:'foo', 'bar' | upcase}}
 {{ "hello" | downcase | upcase}}
 {{ "HELLO" | downcase | upcase}}
-{% assign test = "xample" | upcase | prepend:'E'%}
+{% assign test = "xample" | upcase | prepend:'E' %}
 {{test}}