For the related issue, see #138 
For the unsquashed commit history see [parser-refactor-unsquashed](https://github.com/Goncalerta/liquid-rust/tree/parser-refactor-unsquashed).

I've been looking at some rust parsing libraries and I found `Pest` to seem really simple to use. So I decided to try it out and do some experiments, resulting in this pull request. 

This is just a draft, however, which means that there is no error handling yet, most tags were disabled (I changed the signature of `ParseTag` and `ParseBlock`), there are a lot of TODO's, issues and other things to improve in the code. 

Something I liked about pest is that most of the  error handling will be isolated from the rest of the parsing code. Also, Pest errors give information about line and column, which is a great help for providing more context to errors.

Implementing tags and blocks is now much easier, as the tokens are now "high level" (as a consequence of changing the signature). I made structs (`TagToken` and `TagBlock`) that replace `&[Token]` and `&[Element]` of the old parser, respectively. Instead of working with "low level" tokens, such as `Token::Pipe` or `Token::OpenSquare`, TagToken works with more ready to use ones, such as `Range` or `FilterChain`. TagBlock allows to easily parse everything inside a block, or to just get its raw content.

About performance, from what I've read, `Pest` is slower than `nom` (which is a natural tradeoff of its easiness of use). Still, I hope it's fast enough for this crate's purposes.

@epage Do you like this idea? Should this go forward or is it better to stick to regex/use something like `nom`?

## Features

- Rewrite the whole parser using [Pest](https://pest.rs/);

## Issues solved

- #138 
- #145 
- (Partially) #232
- #226 
- #227
- #242

## Breaking changes

Behavior
- Expressions no longer support tags (they weren't supposed to)
- More strictness in tokens accepted (Tags will raise an error if given a surplus of arguments. This is to alert the user for possible mistakes);

API
- Changed signature for tags and blocks.
- `compiler::parse` takes a`&str` directly as an argument, instead of requiring the midstep of `tokenize`.