@epage 2018-11-06
Good work on this and good idea to do a vertical slice of this to judge feasibility and to get feedback.

Looking at it from the prospect of our goals
- Performance
  - Doesn't seem to do as many passes over the file as we currently do
  - Doesn't have extra allocations in the Token before moving to Renderables (I assume, its all internal to Pest)
- Make it easier to implement new features: Sure looks like it
- Easier to maintain: yup
- Concise top-level that reduces churn
  - This is my biggest area of concern
  - Is Pest in our public API? I haven't dug too much into these `TagToken` and `TagBlock`s
    - To be clear, I'm fine with having parsing utilitiies that live in a sub-crate that we can break compatibility on
    - I just don't want parser changes to break compatibility on Tags / Blocks and the overall top-level API
  - Can a third party add new tags / blocks without requiring updating our parser?
- File location information: Sounds like you've got it!
  - Is this provided in a way that we can update our filters, tags, and block implementations to track their location so semantic errors (and not just syntactic) can provide it?

I'll do a scan through and comment on details as well.