@epage 2018-11-06
Not finding it but I thought I remember seeing a comment about preserving `{{ value }}` being treated differently than `{{ value | filter }}`.  I don't know if I'd worry about that myself.