@epage 2018-11-06
I think I'm done with my high level pass.  Please feel free to express your thoughts; don't take me being unsure about the compatibility aspects as a go/no-go on this.