@epage 2018-11-06
RE Keeping our API stable

Looking at pest, it seems reasonable for us to be able t find a way to keep it out of our API and allow changing parsing engines without a problem.  

The big question is whether to try to do all of the changes in a single PR or for us to do it over time.  A single PR means we touch all of the tags / blocks in one go but I also find its harder to design and iterate when doing that.  Probably better to do multiple passes.

So after reading my comments, if you feel this is the best way moving forward, go for it! I have nothing against it.