@Goncalerta 2018-11-06
>   * Is Pest in our public API? 

I don't think so, at least not after this pull becomes complete. The parser's public API is going to be `parse(text: &str, options: &LiquidOptions) -> Result<Vec<Box<Renderable>>>`, which does not leak anything about the parser itself, `TagBlock` and `TagTokens`, which were created exactly to encapsulate Pest's `Pairs`, providing useful methods to easily implement new tags/blocks.

>     * I just don't want parser changes to break compatibility on Tags / Blocks and the overall top-level API

Once this PR is done (which will inevitably break the compatibility on Tags/Blocks) I hope no more big compatibility breaks will occur. That could only happen, I think, if `TagBlock` and `TagTokens` needed to change their API and I hope they are stabilized once this is finished. 

>   * Can a third party add new tags / blocks without requiring updating our parser?

Well, right now, this draft is very limited. But my goal is to make `TagBlock` and `TagTokens` flexible enough to allow this.

> * File location information: Sounds like you've got it!
>   
>   * Is this provided in a way that we can update our filters, tags, and block implementations to track their location so semantic errors (and not just syntactic) can provide it?

I still have to think about it, as I haven't put much thought in the error handling yet. Semantic errors would be to, for example, try to access a nonexisting variable, right? That can only be known during rendering, as it depends on the context.

As such, the only way I could see this happening is to update structs like `Variable` to store their location. Then, when creating `Variable`s, the parser would pass in their location. Once an error happens, the error crate could look into that information and produce an error message.