@Goncalerta 2018-11-06
> Not finding it but I thought I remember seeing a comment about preserving `{{ value }}` being treated differently than `{{ value | filter }}`. I don't know if I'd worry about that myself.

Probably you're refering to the `grammar.pest` comments. That was a suggestion I made in an earlier phase so `Value` and `Filterchain` would be parsed differently on TagToken, so one wouldn't "eat" the other when parsing. But right now `Filterchain` "eats" everything and `TagTokens` takes care of "unwrapping" the value out of the filterchain so I think there won't be a problem anymore.