@epage 2018-11-06
> Well, right now, this draft is very limited. But my goal is to make TagBlock and TagTokens flexible enough to allow this.

I'm looking forward to your thoughts on how this will be possible.  Like I said in a follow up, I'm not bothered if it doesn't work out because I think we'll still have a path forward for it.