@epage 2018-11-06
> What else must be taken into account to achieve the desired flexibility?

What if all tags / block knowledge is third party, how dwell can our parser handle that?