@Goncalerta 2018-11-06
> > What else must be taken into account to achieve the desired flexibility?
> 
> What if all tags / block knowledge is third party, how dwell can our parser handle that?

In that extreme case I guess it could have a function like `parse_next_element_as_str` which would yeld the next element without parsing it (it would return, for example, "{% tagname something blahblahblah %}"). And the user could make the parsing on his own.

TagTokens could similarly have a `to_str` that would yeld whatever wasn't parsed yet as a &str.

This is the only way I'm seeing if we can't really do any assumptions at all regarding how it would be parsed. But this would also mean the user would be responsible for keeping track of line and column numbers in case they want to raise a semantic error during their custom parse.