@epage 2018-11-06
> In that extreme case I guess it could have a function like parse_next_element_as_str which would yeld the next element without parsing it (it would return, for example, "{% tagname something blahblahblah %}"). And the user could make the parsing on his own.

So again, I don't think we need to solve this in the first iteration but I want us to be able to eventually support new, unthought of tags/blocks without having to modify liquid-rs.

When we get there, another approach is we remove SOI / EOI and parse up until the next tag / block.  We then hand over where we left off in the `str` to the tag/block (plus line/column info).  Once it reaches the end of the block, it returns where it left off in the str and we start parsing again.