@epage 2018-11-14
Again, thanks for looking into this.  There are a lot of tricky or annoying cases this requires and you are doing a great job staying on top of it all.