@epage 2018-11-19
btw opened #232  to help track some of the less obvious requirements we'll eventually need to handle for reporting line numbers in errors.  I don't think its needed in this PR but we should get to it soon after.