@epage 2018-11-23
FYI I'm working on porting integration tests over from the ruby implementation to be able to act as a conformance test.  Once we get the improved parser in, there will be lots of opportunities to tweak it to behave more correctly :)