@Goncalerta 2018-11-26
On `benches/handlebars-liquid.rs` line 9 we have:
```