@Goncalerta 2018-11-26
On `benches/handlebars-liquid.rs` line 9 we have:
```
static ITERATE: &'static str = "<html>
  <head>
    <title>{{year}}</title>
  </head>
  <body>
    <h1>CSL {{year}}</h1>
    <ul>
    {% for team in teams %}
      <li class=\"{% if forloop.index0 == 0 %}champion{{% endif %}\">
      <b>{{team.name}}</b>: {{team.score}}
      </li>
    {% endfor %}
    </ul>
  </body>
</html>";
```

Notice that part `champion{{% endif %}` is that extra `{` before the tag supposed to be there? Is liquid supposed to accept that as a raw instead of raising an error? Because right now in this PR it doesn't and that's causing travis to fail.