@epage 2018-11-26
Performance:
- `parse_template` went from 42ms to 33ms
- `parse_text went from 592ns  to 1654ns
- `bench_parsing_basic_template` went from 31ms to 27ms
- `bench_big_loop_big_object` went from 866ms to 967ms

Source
- https://travis-ci.org/cobalt-org/liquid-rust/jobs/459806284
- https://travis-ci.org/cobalt-org/liquid-rust/jobs/459831175

Caveat: this is relying on Travis which is not very predictable in its performance.